backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.35.0",
    "motor>=3.3.1",
    "python-dotenv>=1.0.1",
    "requests>=2.31.0",
    "opencv-python-headless>=4.10.0",
    "numpy>=1.26",
    "Pillow>=10.3.0",
    "python-multipart>=0.0.6",
    "websockets>=12.0",
    "httpx>=0.27.0",
    "google-auth>=2.23.4",
    "google-auth-oauthlib>=1.1.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]
[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.6.4",
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]

[project.scripts]
start = "uvicorn:main"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "N", "B", "A", "C4", "T10", "ISC", "TCH"]
